docker-container/finlab python/apps/dashboard-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
httpx==0.25.2
python-multipart==0.0.6
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3
python-dateutil==2.8.2
pytz==2023.3
